requirements.txt
# 한국어 금융교재 퀴즈 생성기 v2.0.0 Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# AI & ML
openai==1.45.1
sentence-transformers==2.2.2
torch>=2.6
numpy<2
transformers>=4.21.0

# Vector Database
chromadb==0.4.18

# Data Processing
python-dotenv==1.0.0
pydantic==2.5.0

# Logging
loguru==0.7.2

# Development
langchain==0.1.0 